curando el retraso

+++ HitosIV_test.go
@@ -22,7 +22,7 @@ func TestTodosHitos (t *testing.T){
 
 func TestUno(t *testing.T){
 	t.Log("Test Uno");
-	for i:=0; i < CuantosHitos(); i++{
+	for i:=0; i < (int)CuantosHitos(); i++{
 		t.Log(i);
 	}
 }
